Renamed OgreWidget files to match the class name
