+++ qtEditor/OgreWidget.h
